+++ server/tests/fixtures/database_fixtures.py
@@ -115,6 +115,12 @@ async def flow_run_id(flow_id):
 
 
 @pytest.fixture
+async def running_flow_run_id(flow_run_id):
+    await api.states.set_flow_run_state(flow_run_id, state=Running())
+    return flow_run_id
+
+
+@pytest.fixture
 async def labeled_flow_run_id(labeled_flow_id):
     return await api.runs.create_flow_run(flow_id=labeled_flow_id, parameters=dict(x=1))